backend/requirements.txt
fastapi
uvicorn[standard]
motor
pymongo
python-jose[cryptography]
bcrypt
python-dotenv
pydantic[email]
slowapi
python-multipart
starlette
motor-types

